Changes to inference rules (3)
Only create the inference rule for double colon rules once, not
for each rule without commands.

For single colon rules use a separate variable to store the commands
to be run rather than add it to a new rule.  It's only necessary
to add a new rule to handle the implicit prerequisite, so not for a
.DEFAULT rule.