Add a testsuite
Commit c5869fe (Add a function to find commands for a target) was
a bit of a disaster, causing two serious regressions.

Being firmly in favour of post-horse-bolting of stable doors I've
added a testsuite that would have detected those regressions.
Of course, the trick would be to detect the *next* ill-advised
commit.

The testsuite is based on the one from BusyBox because:

- I'm familiar with it;
- I'm lazy.

As a result you get another copy of GPLv2 to add to your collection.
Yay!