Skip duplicate entries in $? and $^
The GNU make documentation says of $^:

   if you list a prerequisite more than once for a target, the
   value of $^ contains just one copy of the name.

Although it doesn't says so, this also applies to $?.  Both BSD
make and SunPro make do the same, though neither mentions this in
their documentation.  There's also no mention of this feature in
POSIX.

Since skipping of duplicate entries seems to be widely supported
implement it as an extension.