Add support for .PHONY special target
Prerequisites of the .PHONY special target are themselves targets
which are considered to be out-of-date when make is run.  They
aren't affected by the presence of a matching file, nor do they
affect such a file.

This makes the traditional use of targets like 'clean' more reliable
because they won't be upset by the accidental presence of a file
called 'clean'.

The .PHONY special target may appear in a future version of POSIX:

   https://www.austingroupbugs.net/view.php?id=523

(GitHub PR #14)