Append to immediate-expansion macros in POSIX 202X
The code to handle immediate-expansion macros in setmacro()
was conditional on non-POSIX extensions being enabled.  This is
insufficient: '::=' and '+=' are POSIX 202X features so immediate-
expansion macros should also work if they are enabled.

Prior to this commit the test 'Appending to immediate-expansion
macro' failed when make was compiled with extensions disabled
and POSIX 202X features enabled.