Tweak handling of whitespace on input
The input code was somewhat random in its use of isblank() or
isspace() to skip whitespace characters.  In a number of places
POSIX is quite specific that only blanks are to be skipped (that
is, tabs or spaces).  Be more rigorous in applying this.

Also, both process_line() and process_command() now terminate the
line being processed at a non-escaped newline.  This avoids the need
for additional checks for newline.