Allow / and - in target names
Austin Group defect report 1615 has been accepted.  This adds
slash and hyphen to the list of characters required to be
supported in target names.

Add this support as a POSIX 202X feature.

A minor issue with support for hyphen in macro names has also been
fixed.  This didn't affect the default configuration.

(GitHub issue #17)