Show location of errors during build
Error messages only displayed the location of an error while the
makefile was being parsed.

Store the location (makefile name and line number) of each command
line.  This allows location information to be displayed in errors
encountered while targets are being built.

Take this Makefile, for example:

target:
	@echo $(hello

Previously this would display:

   make: unterminated variable '$(hello'

Now it shows:

   make: (Makefile:2): unterminated variable '$(hello'