Add .PRAGMA special target
Add the special target .PRAGMA.  This allows certain features
of the POSIX standard to be disabled when POSIX mode is being
enforced.  This is done by setting prerequisites of .PRAGMA,
as follows:

- macro names containing '-' are permitted with the 'macro_name'
  prerequisite;

- target names containing '-' or '/' are permitted with the
  'target_name' prerequisite;

- the '#' character isn't treated as the start of a comment in a
  command line with the 'command_comment' prerequisite;

- an empty suffix is permitted in a macro expansion of the form
  $(VAR:=.c) with the 'empty_suffix' prerequisite.

Additionally the 'posix_202x' prerequisite enforces the future
POSIX standard instead of the current one.  In this case the
'macro_name' and 'target_name' pragmas aren't required as the
future standard will allow the additional characters.

.PRAGMA without any prerequisites restores the default behaviour.