Avoid superfluous warning messages
pdpmake was rather too prolific in generating 'nothing to be done
for XXX' messages.  GNU make only issues the equivalent message
for top-level targets and only if nothing at all was (or would have
been) done.  bmake doesn't seem to have such a message.

Follow the practice of GNU make.  This requires more intensive
tracking of the actions taken while the make is in progress.

See:

  https://github.com/rmyorston/busybox-w32/issues/354#issuecomment-1692410640