Return non-zero exit status when a command fails
When a build command returned a non-zero exit status pdpmake
reported a warning and returned an exit code of zero.

This was due to the misuse of the status returned by system(3).
As the man page says:

   the return value is a "wait status" that can be examined using
   the macros described in waitpid(2).  (i.e., WIFEXITED(),
   WEXITSTATUS(), and so on).

Use the error() function to correctly report the problem on stderr
and return an exit status of 2.

Some additional changes in the same area:

- When a target is removed report the diagnostic on stderr, as
  required by POSIX.

- When a build command receives a signal GNU make removes the target.
  bmake doesn't and it isn't required by POSIX.  Implement this as an
  extension.

- Expand the error message when a build command fails so it includes
  the exit status or signal number, as obtained from the value
  returned by system(3).