Flush stdout after writing command string to it
Otherwise, if stdout is fully buffered (e.g. because the output is
redirected to a file or pipe) then the command string may still
be buffered when the command is executed in which case the command
string will be written to stdout **after** the command has written
its output to stdout.