Change how macros are read from the environment
Commit 3cb1747 (Enforce restrictions on macro and target names
in POSIX mode) skipped validation of macro names imported from
the environment.  However, it failed to update process_macros()
to suit.  Thus, for example, SHELL was imported when it shouldn't
have been.

The problem this was supposed to address was that of environment
variables with invalid names, which the user may not have control
over, causing a fatal error.

As an alternative, silently ignore variables with invalid names
on import from the environment.