Reset $< and $* if necessary
POSIX only requires $< and $* to be set for inference rules, not
target rules.  So, of course, pdpmake didn't set them for target
rules.

However, this could result in a target rule seeing stale values
for these two internal macros.  If they haven't been given an
explicit value set them to an empty string.