Add posix_2017 pragma
The pragma 'posix_202x' causes the 202X POSIX standard to be
enforced in POSIX mode.  Add an equivalent 'posix_2017' for the
2017 standard.

The way the default standard is configured at build time has been
changed.  There's now a separate DEFAULT_POSIX_LEVEL preprocessor
symbol.  This defaults to the 2017 standard but it can also be
set to 202X.