Fixes to -t option
The -t option (in general) causes targets to be touched instead of
having build commands run to create them.  There were two problems.

The flag variable 'ssilent' in docmds was too small (uint8_t) to
contain the value of 'dotouch' (uint32_t).  Truncation of the
value resulted in build commands being echoed when they shouldn't
have been.

The POSIX specification is unclear as to how build commands with
a '+' prefix interact with touch.  The rationale indicates that
this feature was imported from GNU make, so the behaviour has been
made to match what it does:  if a '+' build command is run the
target is not touched.

The code has been rearranged to move the call to touch() up into
docmds().