Fixes to -q option
The -q option returns an exit status to indicate if targets are
up-to-date (0) or in need of updating (1) but without updating
them.  As an exception (imported from GNU make) build commands
with a '+' prefix are executed.

pdpmake didn't implement the exception.  Doing so required moving
handling of the -q option down into docmds().