Add -h option to print usage message
The man page says that 'pdpmake -h' displays help information.

Update the code to agree with that.  Although this is a non-POSIX
extension it conforms to the POSIX requirement that non-error
output should go to stdout.

Update the man page to include 'h' in the list of options.

Update the Makefile to add the project and volume names, as in the
original man page.