Restore warning about invalid macro name
Commit 7591b1b (Change how macros are read from the environment)
was intended to ignore environment variables with invalid names.
It had the unintended consequence of also ignoring macros with
invalid names defined in makefiles.  This was because such macros
can have the same level (3) as those imported from the environment.

Rather than use the level to detect importing from the environment
add a flag to indicate this circumstance.