Explicitly verify order of arguments
POSIX requires macro assignments to precede targets on the command
line.  This requirement has been relaxed as a non-POSIX extension.

Verify the constraint has been applied in POSIX mode.  This ensures
that the command:

   make --posix target A=1

fails before it tries to update the target.

It also catches the case where POSIX mode had not been set with the
--posix option or PDPMAKE_POSIXLY_CORRECT environment but was then
set by the .POSIX special target in the makefile.  In this case the
command line was scanned for macro assignments without enforcing
POSIX mode while targets were processed with POSIX mode enabled.