Duplicate makefile name recorded with command
Commit 4f0be50 (Show location of errors during build) stored a
pointer to the name of the makefile with each command, for use
in diagnostic messages.

While this is fine for makefiles defined on the command line, the
names of included files will have been freed before they can be
used.

Always take a copy of the makefile name stored with commands.

(busybox-w32 GitHub issue 449)