Reinstate conditional skipping of command lines
Commit e422a86 (Allow empty commands) rearranged readline() in
a way that broke the use of conditionals within the definition
of a rule.

Add a test case to detect this.  Adjust readline() so that
conditionals are processed before returning command lines or
checking for empty lines and comments.

Remove the test for a leading tab in skip_lines().  This allows
conditionals in the definition of a rule to be indented with a
leading tab.