Fix test for include with no pathnames
In POSIX 2024 the behaviour of an include line with no pathnames
is unspecified.  The test for this condition was incorrect in
certain circumstances.

Files containing dependencies may be generated by the compiler and
included in the makefile.  On an initial build no dependency files
will be present and the include line is made to ignore errors.  In
this case the test for no pathnames returned true, even though many
pathnames may have been present.

Fix the problem by always setting 'makefile' even if the named
file can't be opened.