Fix error reporting for included files
The global variables 'makefile' and 'lineno' weren't properly
reset after a recursive call to input().  As a result error
messages for included files could refer to the wrong file or
line number.

Restore 'makefile' and 'lineno' immediately after a rescursive
call to input().  This also requires a different fix for commit
4186ebb (Fix test for include with no pathnames).