Changes to .DEFAULT/inference rules
The POSIX standard allows inference rules to be redefined but not
the .DEFAULT rule.  There is no explicit exception for .DEFAULT to:

   Only one target rule for any given target can contain commands.

Treat redefinition of a .DEFAULT rule as an error in POSIX mode
but allow it as an extension.

Also, the code didn't allow an inference rule with dependencies to
redefine an existing inference rule.  This is no longer the case.