Phony targets and double colon rules
Commit 6fd5e72 (Disallow inference rules for phony targets)
added the GNU/BSD make feature that inference rules aren't used
for phony targets.

Normally a double-colon rule without associated commands causes
an inference rule search.  If the target of the double-colon rule
was phony no search was carried out and the rule failed.  As a
result of this perl failed to build.

When a double-colon rule has no commands and a phony target the
prerequisites should be built and the rule should succeed.

(GitHub issue #70)