Fix detection of target rule with inline command
Commit 32bf27d (Fix detection of target rules (take 2)) added code
to handle the case where a target rule had an inline command with
a ';' separator in a macro.  This required all macros on the line
to be expanded, including those in the inline command.

The ';' should first be searched for in the original line (without
any macro expansion) and only if that fails should macros be
expanded.

This matches the behaviour of GNU and Unix V7 make.  BSD and Schily
make don't handle the case where the ';' is in a macro.

(GitHub issue #73)